Added configuration command.
The configuration command adds the configuration dialog given by
KTextEditor. The dialog saves the configuration settings automatically
and automatically restores the settings. I am not exactly sure how
KTextEditor saves / loads the settings; it seems that the settings are
saved in ~/.config/FreeCADrc. So far it seems to work correctly so I
will let KTextEditor to handle the configuration.